cpu_input_boost: Remove scheduler boost on screen wake
Forcing a large migration of tasks to the big CPUs is not good to do when
the screen wakes, as it adds a lot of additional latency. This results in
everything feeling laggy when waking the screen sometimes.

Remove the scheduler boost from the screen-wake boost in order to reduce
system latency when the screen wakes.

Signed-off-by: Sultanxda <sultanxda@gmail.com>
Signed-off-by: Ahmad Thoriq Najahi <tlogitechnjhi@gmail.com>